refactor: extract FeatureStreamEndpoints partials per endpoint (#1144)
FeatureStreamEndpoints.cs was a 2,544-LOC god-file. Split the class
across four partial files by responsibility, keeping the public
surface (MapFeatureStreamEndpoints + admin endpoints + capabilities)
intact in the root file:

  - FeatureStreamEndpoints.SubscriptionFilter.cs (574 LOC):
    query-string parsing/validation and shared layer authorization
    helpers.
  - FeatureStreamEndpoints.WebSocket.cs (1,248 LOC): WebSocket
    transport, channel-drain writer, control-frame dispatcher,
    subscribe/unsubscribe handlers, and framing/dedup helpers.
  - FeatureStreamEndpoints.Sse.cs (382 LOC): SSE transport
    handshake, replay/handoff, and channel-drain loop.
  - FeatureStreamEndpoints.Replay.cs (145 LOC): durable-store
    replay helpers shared by both transports.

The root file is now 276 LOC and only owns endpoint mapping,
transport dispatch, capabilities, admin endpoints, and small
shared helpers. No behavior change.

Build clean. Feature-stream test counts unchanged from pre-refactor
baseline (92 passed / 19 failed; the 19 failures are pre-existing
on this branch and reproduce against the unrefactored file).
