+++ src/Honua.Server/Features/Streaming/FeatureStreamEndpoints.Replay.cs
@@ -1,0 +1,145 @@
+// Copyright (c) Honua. All rights reserved.
+// Licensed under the Elastic License 2.0. See LICENSE in the project root.
+
+using System.Net.WebSockets;
+using System.Text.Json;
+using Honua.Server.Features.Infrastructure.Events;
+
+namespace Honua.Server.Features.Streaming;
+
+/// <summary>
+/// Durable-store replay helpers shared by the WebSocket and SSE transports.
+/// Walks the feature-change event store from a cursor, applies subscription
+/// filters and per-(event, subscription) dedup, then writes envelopes to the
+/// transport.
+/// </summary>
+internal static partial class FeatureStreamEndpoints
+{
+    private static async Task<long> ReplayToWebSocketAsync(
+        WebSocket webSocket,
+        SemaphoreSlim writeLock,
+        IFeatureChangeEventStore eventStore,
+        long fromCursor,
+        int batchSize,
+        ILogger logger,
+        Guid sessionId,
+        CancellationToken cancellationToken,
+        IStreamSubscriptionFilter? subscriptionFilter = null,
+        string? subscriptionId = null,
+        FeatureStreamSessionManager? sessionManager = null,
+        long subscriptionGeneration = 0)
+    {
+        var cursor = fromCursor;
+        while (!cancellationToken.IsCancellationRequested)
+        {
+            var events = await eventStore.QueryAsync(cursor, null, null, batchSize, cancellationToken).ConfigureAwait(false);
+            if (events.Count == 0)
+            {
+                break;
+            }
+
+            FeatureStreamLog.ReplayStarted(logger, events.Count, cursor, sessionId);
+
+            foreach (var evt in events)
+            {
+                var envelope = FeatureStreamPublisher.ToEnvelope(evt) with { SubscriptionId = subscriptionId };
+                cursor = evt.Cursor;
+
+                // Apply subscription filter during replay — advance cursor past filtered events.
+                if (subscriptionFilter is not null
+                    && !subscriptionFilter.Matches(envelope, evt.GeometryEnvelope, evt.PropertiesJson))
+                {
+                    continue;
+                }
+
+                // When a session manager and generation are supplied, claim the
+                // (event, subscription) slot atomically. The claim also verifies
+                // the subscription's generation, fencing stale replays after an
+                // unsubscribe/replacement (although the per-connection control
+                // loop is single-threaded, so this matches the writer-side
+                // contract). Whichever send-time path wins the atomic test-and-
+                // set sends the frame; the other observes the recorded key and
+                // skips.
+                if (sessionManager is not null && subscriptionId is not null && subscriptionGeneration > 0)
+                {
+                    if (sessionManager.TryClaimSubscriptionDelivery(sessionId, subscriptionId, subscriptionGeneration, evt.EventId)
+                        != SubscriptionDeliveryClaim.Claimed)
+                    {
+                        continue;
+                    }
+                }
+                else if (sessionManager is not null && subscriptionId is not null)
+                {
+                    // Generation-less call site (legacy/test); fall back to the dedup-only path.
+                    if (!sessionManager.TryRememberSubscriptionDelivery(sessionId, subscriptionId, evt.EventId))
+                    {
+                        continue;
+                    }
+                }
+
+                var payload = JsonSerializer.SerializeToUtf8Bytes(envelope, FeatureStreamJsonContext.Default.FeatureStreamEnvelope);
+                await SendWebSocketJsonAsync(webSocket, writeLock, payload, cancellationToken).ConfigureAwait(false);
+            }
+
+            if (events.Count < batchSize)
+            {
+                break;
+            }
+        }
+
+        return cursor;
+    }
+
+    private static async Task<long> ReplayToSseAsync(
+        HttpResponse response,
+        IFeatureChangeEventStore eventStore,
+        long fromCursor,
+        int batchSize,
+        ILogger logger,
+        Guid sessionId,
+        CancellationToken cancellationToken,
+        IStreamSubscriptionFilter? subscriptionFilter = null,
+        string? subscriptionId = null)
+    {
+        var cursor = fromCursor;
+        while (!cancellationToken.IsCancellationRequested)
+        {
+            var events = await eventStore.QueryAsync(cursor, null, null, batchSize, cancellationToken).ConfigureAwait(false);
+            if (events.Count == 0)
+            {
+                break;
+            }
+
+            FeatureStreamLog.ReplayStarted(logger, events.Count, cursor, sessionId);
+
+            foreach (var evt in events)
+            {
+                var envelope = FeatureStreamPublisher.ToEnvelope(evt) with { SubscriptionId = subscriptionId };
+                cursor = evt.Cursor;
+
+                // Apply subscription filter during replay — advance cursor past filtered events.
+                if (subscriptionFilter is not null
+                    && !subscriptionFilter.Matches(envelope, evt.GeometryEnvelope, evt.PropertiesJson))
+                {
+                    continue;
+                }
+
+                await WriteSseEventAsync(
+                    response,
+                    "feature-change",
+                    envelope,
+                    FeatureStreamJsonContext.Default.FeatureStreamEnvelope,
+                    envelope.Cursor,
+                    cancellationToken).ConfigureAwait(false);
+                await response.Body.FlushAsync(cancellationToken).ConfigureAwait(false);
+            }
+
+            if (events.Count < batchSize)
+            {
+                break;
+            }
+        }
+
+        return cursor;
+    }
+}
